chore: Move `FileUpload` to model.ts [skip pizza]
